Cleanup promexporter
Change NewMetrics to take a ebpf.Map and ebpf.MapSpec instead of a
"loader", eliminate "loader". It was done to simplify usage.

Loader was conceived as a tool allowing complex *modular* eBPF app
initialization. Consider a module that needs *both* to augment the spec
and grab references to eBPF objects. It would require multi-step module
initialization. The first step would "fix" the spec while the subsequent
one would grab references. It is not possible to combine two steps into
one since we might need to apply further spec rewrites before loading.
Loader was an attempt to enable one-step initialization via incremental
loading.

Promexporter doesn't augment the spec hence it doesn't really need a
"loader". In fact, we didn't have a compelling case for the "loader" so
far.

Signed-off-by: Nick Zavaritsky <mejedi@gmail.com>